pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sumula"
version = "0.1.0"
description = "Construtor de Súmula Curricular FAPESP"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "arq>=0.25",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "pydantic-settings>=2.0",
  "python-multipart>=0.0.9",
  "jinja2>=3.1",
  "slowapi>=0.1.9",
  "openai>=1.30",
  "pymupdf>=1.24",
  "pandas>=2.2",
  "openpyxl>=3.1",
  "trafilatura>=1.12",
  "bibtexparser>=1.4",
  "httpx>=0.27",
  "python-magic>=0.4",
  "lxml-html-clean>=0.2",
  "scholarly>=1.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
